chore: move `LearningRateExp` to `deepmd.utils.learning_rate`
